Rename DiscoverByGenreItem to ShowSummaryView
- first step in an effort to reduce the number of custom views, working
from a bottom up approach to see how far it can go without complicating
the Activity like it used to be in days of old. By reducing the number
of custom views, it's hoped to make maintenance (of naming conventions)
easier, and on boarding easier too
- Note the view is package protected (ctors must stay public for system)
so it's not qualified with the screen name (Discover) but any external
association with this view is associated with screen name (merge layout,
styles, IDs)
- The view layout reverted to being named after where it's
used rather than what's inside. This lowers the number of resources
coupled with the custom view, hopefully making maintenance easier
- The view is reverted to being pretty dumb - it doesn't know about the
model, it's got view-specific things like setTitle and setPoster, so if
the model changes, the ViewHolder will be updated, but not the view
(unless it needs to show different information)