Fix (I think?) sometimes not getting a full set of ShowsInGenre
- sometimes, e.g. onResume occasionally, Discover will show only a
subset of ShowsInGenre, i.e. not all the genres.
- I think this is because previously, there was a race condition with
`configurationObservable()` and `discoverTvShowsObservable`, where an
onCompleted event would occur in one of these and then the combineLatest
would just stop?
- Changing to a zip _should_ have no real impact - it's the repeat() on
the configurationObservable which I think will help; now, the only thing
that will have an onCompleted event is the discoverTvShowsObservable.
- Alternatively, could have left the combineLatest and lifted the
`configrationObservable` to swallow onComplete events and I think same
effect